Add MQTT support to climate.py
This commit adds support for an MQTT controller to the climate component.
By specifying controller_send_service of mqtt.publish in configuration.yaml
along with supportedController of MQTT in the code file, climate commands
will be published to the topic defined in command_topic.
The included 1520.json file provides hex codes (tested and verified) for
Kelvinator A/C using a R51K/BGE remote with coolix infrared protocol.