Add support for Trotec PAC 2010 E
This is a rebranded AC from GREE, possibly also works for other AC's. I reckon it works for everything that uses the GREE protocol.

This is made to be used with MQTT and Tasmota on a sonoff RF bridge! Warning: TASMOTA will have to be modified to support the GREE codes properly! Once that requirement is fulfilled this works like a charm.