Added play_media function to allow channel change functionality useful with alexa, google or from the homeassistant app as well as play,pause.
Channels can be selected either by a channel number or by a channel name which is looked up in the channels list to retrieve the channel number.

Added attribute "volume_step_default" that can be used by code in Alexa to set the default volume step increment when an Alexa volume up command is received. The current Alexa default is 10 which is too high.

The Alexa smart home code was also modified to add stepSpeaker function to allow volume to be set in variable increments.