Add codes for Whynter portable air conditioners.
This was tested with a Whynter ARC-126MDB, but should work with other
types of Whynter air conditioners. It uses Fahrenheit units, but can
easily be converted to Celsius - just replace 62 with 17, 64 with 18,
and so on, until 88 with 30. (That's different from the actual
temperature conversions, but lines up with what the display will read.)