Add codes for Lasko FH500.
The Lasko FH500 is a combination tower fan/space heater. It has four fan
modes, three heat modes, and a 'heater fan' mode that enables the heater
and the fan at the same time. It also has an 'Auto Eco' mode that
enables a built-in thermostat.

This adds support for the FH500 as follows:
- In fan-only mode, fan modes 'low', 'medium', 'high', and 'max' are the
four fan modes, in increasing order of fan speed. The target temperature
is ignored.
- In heat mode, fan modes 'low', 'medium', and 'high' are the three heat
modes (650W, 1000W, and 1500W, respectively), and 'max' enables the
'heater fan' mode (highest power heat + lowest speed fan). The target
temperature is ignored.
- Using auto mode enables Auto Eco mode. The fan mode is ignored, and
the target temperature is passed on to the FH500.
- The target temperature supports half-degree steps. The fractional part
of the target temperature controls the oscillation mode: X.0 degrees
turns off oscillation, and X.5 degrees turns it on (setting the FH500 to
X degrees, if in auto mode).