climate: Support non-standard precision in climate entities.
Ordinarily, precision is based on the temperature unit - Fahrenheit
users see a precision of 1F, while Celsius users see a precision of
0.1C. If the temperature step of a climate device is smaller than this,
it results in display issues as Home Assistant allows you to set the
more-precise temperature, but then rounds it for display and storage.

This change will make the precision more fine-grained, if necessary to
be able to represent a smaller temperature step.