Update CLIMATE.md
Added "Aire+" , which is a white label of "Hisense" sold in Europe. The DG11L1-04 remote is, from quick testing I just completed, compatible with the 1522 code file, so it's the one linked.

Alternatively, the Hisense entry could be tweaked to include this remote, but it will not be as straightforward for someone looking for compatibility, since you need to know Hisense is the main brand, which is not common knowledge.