Added automatic conversion between Celsius and Fahrenheit.
The target temperature is now stored in the native units of the Home Assistant instance. Before transmitting the IR code this value is converted into the device's temperatureUnits and rounded to the device's precision.

Since rounding is not an issue for the device's min/max temperature, for simplicity they remain in the device's units, and are converted to the units of the HA instance when queried.

For devices that don't specify temperatureUnit, it is auto-detected based on the min/max temperature range.