Update media_player.py
If on the TV to turn on a two-digit (three-digit) channel, you must first press the "..." button, activating the ability to turn on two-digit (three-digit) channels, then to turn on the one-digit channel, you must press the "OK" button after the "..." button. On TVs where pressing the "..." button is not necessary, completing the channel command with the "OK" button will only speed up the switching.