Detachment of the list of commands for selecting a signal source from the list of signal sources.
There is one serious flaw in the current implementation of the media player: if I don't want to see channels in the source list (on TV), then I can't select a channel using a script (with a number).
This happens because channel switching commands are removed from the list of commands using a custom configuration (not everyone needs channel numbers in the list of sources, this is personally inconvenient for me).
Adding a copy of the dictionary with commands for generating a list of signal sources solves this problem, since in this case the original set of commands remains unchanged and switching channels by number will be available. At the same time, the list of devices will be based on the list of commands from the device configuration, but can be shortened (edited) by the user in the yaml device configuration without affecting the functionality of the device.