Saving the device state when shutting down and restarting Home Assistant (yaml setting)
Some (many) devices can save the selected state before turning off and restore it after turning on. This change will also allow you to remotely see what state the device is in: what signal source is selected, the sound mode, and whether the sound is muted. But all this without feedback from the device. On the other hand, the on / off state may also not have feedback. Therefore, for those who may need this function, they will have to configure their device in yaml