Avoid resync when value is not changing
When using this with Adaptive lighting, during the day the light keeps beeping due to reaching the end of its range every 90s.
This is because Adaptive lighting is trying to adjust the brightness and color temperature, but they are not actually moving because of the limited number of steps.
Putting the resync code inside the condition that we actually need to move the brightness should prevent this, and only resync when we are actually changing to the end of the range.