Update and rename 1000.json to 1002.json
This file is similar to file 1000. It differs from it that it is for Hommer air conditioners intended for the Libyan market, as it does not differ from the previous file except with the feature of turning off the conditioner, which was taken from the same remote control of the Hommer air conditioner.